add LGPL license header
